replaced size_t with uint64_t

+++ include/DocxFactory/DocxCompiler/DocxCompilerBookmark.h
@@ -26,20 +26,20 @@ XERCES_CPP_NAMESPACE_END
                 DocxCompilerItemFile* p_itemFile,
                 DocxCompilerBookmark* p_parent,
                 const string& p_name,
-                size_t p_id,
+                uint64_t p_id,
                 xercesc::DOMElement* p_startNode);
 
         virtual ~DocxCompilerBookmark();
 
         void normalize();
 
-        void setGroupId(size_t p_groupId);
+        void setGroupId(uint64_t p_groupId);
         void setEndNode(xercesc::DOMElement* p_node);
         void insertChild(DocxCompilerBookmark* p_bookmark);
 
         string getName() const;
-        size_t getId() const;
-        size_t getGroupId() const;
+        uint64_t getId() const;
+        uint64_t getGroupId() const;
         xercesc::DOMElement* getStartNode() const;
         xercesc::DOMElement* getEndNode() const;
         DocxCompilerBookmark* getParent() const;
@@ -55,8 +55,8 @@ XERCES_CPP_NAMESPACE_END
         DocxCompilerBookmark* m_parent;
         string m_name;
 
-        size_t m_id;
-        size_t m_groupId;
+        uint64_t m_id;
+        uint64_t m_groupId;
 
         xercesc::DOMElement* m_startNode;
         xercesc::DOMElement* m_endNode;
