+++ include/DocxFactory/util/DocxFactoryDefs.h
@@ -45,7 +45,7 @@ namespace DocxFactory {
 
     typedef uint8_t uint8;
     typedef uint16_t uint16;
-    typedef size_t uint32;
+    typedef uint64_t uint32;
     typedef uint64_t uint64;
 
 #endif
